Publish released version and debug version

+++ Final Project/MediaPlayer/MediaPlayer/VideoLoader.cpp
@@ -15,7 +15,7 @@ void VideoLoader::SetStartFrameIndex(int startFrameIndex) {
 	this->startFrameIndex = startFrameIndex;
 }
 
-int VideoLoader::GetTotalFrames() {
+qint64 VideoLoader::GetTotalFrames() {
 	return totalFrames;
 }
